Make the pagination component shared to all component to be reusable, and edit its logic a little bit
